Moved escape handler to modal component
